Split most of the functions in the Server.js file to appropriate seperate files. Establish Mongosb connections, Established Apollo server in server side
