+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "node server/server.js",
+    "start": "node server --ignore client",
     "client": "cd client && vite",
     "client:build": "cd client && vite build",
     "client:serve": "cd client && vite preview",
