+++ server/config/connection.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 
-console.log('MongoDB URI:', process.env.MONGODB_URI);
+//console.log('MongoDB URI:', process.env.MONGODB_URI);
 
 mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/googlebooks')
   .then(() => {
